requirements.txt
flask>=3.0.0
rembg>=2.0.69
pillow>=10.0.0
onnxruntime>=1.18.0
